requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
bleach==3.2.1
boto3==1.17.7
botocore==1.20.7
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
Django==3.1.5
django-cors-headers==3.6.0
django-countries==7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
future==0.18.2
geocoder==1.38.1
geoip2==4.1.0
google-auth==1.24.0
idna==2.10
IP2Location==8.5.1
isort==5.7.0
jmespath==0.10.0
keyring==21.8.0
lazy-object-proxy==1.4.*
lxml==4.6.2
Markdown==3.3.3
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.8
Pillow==8.1.0
pip-review==1.1.0
pkginfo==1.6.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7
s3transfer==0.3.4
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.56.0
twine==3.3.0
typed-ast==1.4.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
